:sparkles: 런치스크린 진행시간동안 Statusbar가 나오지 않도록 수정했습니다.
